amcrest_gps_pro-master/requirements.txt
aioredis==1.2.0
amqp==2.3.2
anyjson==0.3.3
appdirs==1.4.3
asgiref==2.3.2
async-timeout==3.0.1
asyncio==3.4.3
attrs==18.2.0
autobahn==18.12.1
Automat==0.7.0
billiard==3.5.0.5
blinker==1.4
braintree==3.51.0
cached-property==1.5.1
celery==4.0.2
certifi==2018.11.29
channels==2.1.6
channels-redis==2.3.2
chardet==3.0.4
Click==7.0
constantly==15.1.0
daphne==2.2.4
datedelta==1.2
defusedxml==0.5.0
diff-match-patch==20181111
Django==2.1.4
django-background-tasks==1.2.0
django-common-helpers==0.9.2
django-compat==1.0.15
django-cors-headers==2.4.0
django-cron==0.5.1
django-enumchoicefield==1.0.2
django-import-export==1.2.0
django-mysql==2.4.1
django-rest-framework-mongoengine==3.3.1
django-uuidfield==0.5.0
djangorestframework==3.9.0
djangorestframework-jwt==1.11.0
dnspython==1.16.0
et-xmlfile==1.0.1
geographiclib==1.49
geopy==1.18.1
gpx-parser==0.0.4
gpxpy==1.3.5
h11==0.8.1
hiredis==0.3.1
httptools==0.0.11
hyperlink==18.0.0
idna==2.8
incremental==17.5.0
isodate==0.6.0
jdcal==1.4.1
kombu==4.2.2.post1
lxml==4.3.1
MarkupPy==1.14
mongoengine==0.16.3
msgpack==0.5.6
mysql-connector-python==8.0.13
mysqlclient==1.3.14
netifaces==0.10.6
numpy==1.17.4
odfpy==1.4.0
openpyxl==3.0.0
pandas==0.25.3
protobuf==3.6.1
pusher-push-notifications==1.0.2
pyfcm==1.4.5
PyHamcrest==1.9.0
PyJWT==1.7.1
pymongo==3.7.2
PyMySQL==0.9.3
PySocks==1.7.1
python-dateutil==2.8.1
python-engineio==3.1.2
python-socketio==3.0.0
pytz==2018.7
PyYAML==5.1.2
redis==3.0.1
requests==2.21.0
requests-toolbelt==0.8.0
schedule==0.5.0
six==1.12.0
socketio==0.1.6
suds-py3==1.3.3.0
tablib==0.14.0
togeojsontiles==0.1.2
twilio==6.26.0
Twisted==18.9.0
txaio==18.8.1
typing==3.6.2
urllib3==1.24.1
uvicorn==0.3.24
uvloop==0.14.0
vine==1.2.0
websocket-client==0.54.0
websockets==7.0
xlrd==1.2.0
xlwt==1.3.0
xmltodict==0.12.0
zeep==3.2.0
zope.interface==4.6.0